Fix OPA redirect and SESSION_CREATE
It fixes one bug (1), improves one feature (2) and adds features (3).

(1) The redirect of one-page-application (OPA) subpages was broken.
The commit e68a315 fixes it.

(2) The behaviour of using AFB_SESSION_CREATE wasn't easy to use
and integrate. This was because it enforced the creation of the
session to be the real creation of the session: the first call.
This semantic is now obsolete and replaced with a semantic of
authorisation. The future obsolete AFB_SESSION_CREATE now use
that semantic. The commit 457c879 change this.

(3) The features of closing a session and setting the "level of
authorisation (LOA)" are accessible to plugin as verbs since
commit f3a1fb9.

Change-Id: I02ba84ba661a6813274641b27d114fba6dc4e1c3
Signed-off-by: José Bollo <jose.bollo@iot.bzh>