fix-startup-sequence: remove 4A & BT manager dependencies to bluez-alsa
Since:
- There is no need for bluetooth-manager to require bluez-alsa
- 4A (through the bluez-alsa plugin) does not have strong dependency
on the bluez-alsa service, so that both services can be started
in whatever order (relying on dbus notifications)

... this commit removes the not needed "Requires" on
4a and bluetooth manager.
An annoying effect of keeping it for 4a is that systemd terminates
it whenever bluez-alsa gets restarted. And a coming improvement
will be to have bluez-alsa restart when a HCI device is removed
or inserted.

Signed-off-by: Thierry Bultel <thierry.bultel@iot.bzh>